Move the whole plotting module
