+++ src/einsteinpy/plotting/__init__.py
@@ -1,3 +1,4 @@
 from einsteinpy.plotting.geodesics.core import GeodesicPlotter
 from einsteinpy.plotting.geodesics.interactive import InteractiveGeodesicPlotter
 from einsteinpy.plotting.geodesics.static import StaticGeodesicPlotter
+from einsteinpy.plotting.hypersurface.core import HypersurfacePlotter
